separated the styles from the keyboardTest file
+ added a folder for the keyboardTest file and its styles file
